requirements.txt
click
pyyaml
json-e
